removed code duplication with ugly macros
The duplication of code between the types was annoying and
potentially a maintenance problem. I considered (and tried)
cleaning up with traits, but the current limitations of
traits on Rust stable were a bit daunting. It turned out to
be easier to just spam macros at the problem until it went
away.

This may not be the ultimate solution. Bikeshedding welcome.